added Cosine Similarity
