+++ sentencesimilarities/sentence_similarity.py
@@ -1,10 +1,14 @@
+import sys
 from nltk.stem.snowball import SnowballStemmer
 from nltk.stem import WordNetLemmatizer
 import sentencesimilarities.util.sentence_util as SentenceUtil
 import sentencesimilarities.util.file_util as FileUtil
+import sentencesimilarities.algorithm.symmetric_similarity as Symmetric_Similarity
+import sentencesimilarities.algorithm.cosine_similarity as Cosine_Similarity
 
 stem_obj = SnowballStemmer('english')
 word_net_lemma = WordNetLemmatizer()
+algo_type = "SS"
 
 
 def get_best_suitable_question(user_question):
@@ -15,13 +19,21 @@ def get_best_suitable_question(user_question):
         if score > best_score:
             best_score = score
             best_sentence = db_question
+    print("### Best Score the algo :")
+    print(best_score)
     return best_sentence, best_score
 
 
 def calculate_best_score(db_question, user_question):
     symmetric_similarity_score = 0.0
-    primary_sim = calculate_symmetric_score(db_question, user_question)
-    secondary_sim = calculate_symmetric_score(user_question, db_question)
+    primary_sim = 0
+    secondary_sim = 0
+    if algo_type == "SS":
+        primary_sim = Symmetric_Similarity.calculate_symmetric_score(db_question, user_question)
+        secondary_sim = Symmetric_Similarity.calculate_symmetric_score(user_question, db_question)
+    elif algo_type == "CS":
+        primary_sim = Cosine_Similarity.calculate_symmetric_score(db_question, user_question)
+        secondary_sim = Cosine_Similarity.calculate_symmetric_score(user_question, db_question)
 
     if primary_sim is not None and secondary_sim is not None:
         symmetric_similarity_score = (primary_sim + secondary_sim) / 2
@@ -30,46 +42,6 @@ def calculate_best_score(db_question, user_question):
     elif primary_sim is None and secondary_sim is not None:
         symmetric_similarity_score = secondary_sim
     return symmetric_similarity_score
-
-
-def calculate_symmetric_score(question1, question2):
-    """ compute the sentence similarity using Wordnet """
-
-    # Tokenize and tag
-    sentence1 = SentenceUtil.pos_tagging(question1)
-    sentence2 = SentenceUtil.pos_tagging(question2)
-
-    # Get the synsets for the tagged words
-
-    syn_sets_1 = [SentenceUtil.get_synsets_from_tags(*tagged_word) for tagged_word in sentence1]
-    syn_sets_2 = [SentenceUtil.get_synsets_from_tags(*tagged_word) for tagged_word in sentence2]
-
-    # Filter out the Nones
-    syn_sets_1 = SentenceUtil.remove_empty_sets(syn_sets_1)
-    syn_sets_2 = SentenceUtil.remove_empty_sets(syn_sets_2)
-
-    sim_score, count = 0.0, 0
-    # For each word in the first sentence
-    for syn_set in syn_sets_1:
-        # Get the similarity value of the most similar word in the other sentence
-        best_score = [syn_set.path_similarity(ss) for ss in syn_sets_2]
-        filtered_score = []
-
-        for score_value in best_score:
-            if score_value is not None:
-                filtered_score.append(score_value)
-                # Check that the similarity could have been computed
-        new_score = 0.0
-        if filtered_score:
-            if filtered_score is not None:
-                new_score = max(filtered_score)
-        if new_score is not None:
-            sim_score += new_score
-            count += 1
-    # Average the values
-    if count > 0:
-        sim_score /= count
-    return sim_score
 
 
 def bot_start_chat():
@@ -90,5 +62,6 @@ def bot_start_chat():
 
 
 if __name__ == '__main__':
+    algo_type = sys.argv[1]
     input_question_set = FileUtil.read_questions_from_input_file("resources/FaqQuestions.txt")
     bot_start_chat()
